plat-ti: Move console related functions to a separate file
[Rebase on top of master]
