+++ core/arch/arm/plat-ti/main.c
@@ -31,7 +31,6 @@
 #include <string.h>
 #include <assert.h>
 #include <drivers/gic.h>
-#include <drivers/serial8250_uart.h>
 #include <arm.h>
 #include <kernel/generic_boot.h>
 #include <kernel/panic.h>
@@ -47,7 +46,10 @@
 #include <console.h>
 #include <sm/sm.h>
 
-static void main_fiq(void);
+static void main_fiq(void)
+{
+	panic();
+}
 
 static const struct thread_handlers handlers = {
 	.std_smc = tee_entry_std,
@@ -65,45 +67,6 @@ const struct thread_handlers *generic_boot_get_handlers(void)
 {
 	return &handlers;
 }
-
-static void main_fiq(void)
-{
-	panic();
-}
-
-static vaddr_t console_base(void)
-{
-	static void *va;
-
-	if (cpu_mmu_enabled()) {
-		if (!va)
-			va = phys_to_virt(CONSOLE_UART_BASE, MEM_AREA_IO_NSEC);
-		return (vaddr_t)va;
-	}
-	return CONSOLE_UART_BASE;
-}
-
-void console_init(void)
-{
-	serial8250_uart_init(console_base(), CONSOLE_UART_CLK_IN_HZ,
-			     CONSOLE_BAUDRATE);
-}
-
-void console_putc(int ch)
-{
-	vaddr_t base = console_base();
-
-	if (ch == '\n')
-		serial8250_uart_putc('\r', base);
-	serial8250_uart_putc(ch, base);
-}
-
-void console_flush(void)
-{
-	serial8250_uart_flush_tx_fifo(console_base());
-}
-
-
 
 struct plat_nsec_ctx {
 	uint32_t usr_sp;
@@ -161,5 +124,3 @@ void init_sec_mon(unsigned long nsec_entry)
 	nsec_ctx->mon_lr = plat_ctx->mon_lr;
 	nsec_ctx->mon_spsr = plat_ctx->mon_spsr;
 }
-
-
