+++ core/arch/arm/plat-ti/platform_config.h
@@ -87,12 +87,6 @@
 #define GICC_OFFSET         0x1000
 #define GICD_OFFSET         0x0
 
-#define DEVICE0_PA_BASE		ROUNDDOWN(CONSOLE_UART_BASE, \
-					  CORE_MMU_DEVICE_SIZE)
-#define DEVICE0_VA_BASE		DEVICE0_PA_BASE
-#define DEVICE0_SIZE		CORE_MMU_DEVICE_SIZE
-#define DEVICE0_TYPE		MEM_AREA_IO_NSEC
-
 #define DEVICE1_PA_BASE		ROUNDDOWN(GIC_BASE, CORE_MMU_DEVICE_SIZE)
 #define DEVICE1_VA_BASE		DEVICE1_PA_BASE
 #define DEVICE1_SIZE		CORE_MMU_DEVICE_SIZE
